Code_Files/requirements.txt
cffi==1.15.0
cryptography==37.0.2
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.20.0
six==1.16.0
tabulate==0.8.9
pytest==7.1.2
